PICARD-183 - Add Option to copy files and leave source files unattached

+++ picard/file.py
@@ -157,7 +157,8 @@ class File(LockableObject, Item):
     def _save_and_rename(self, old_filename, metadata, settings):
         """Save the metadata."""
         new_filename = old_filename
-        if not settings["dont_write_tags"]:
+
+        if not settings["dont_write_tags"] and not settings["copy_files"]: # save here only if files get moved,not copied
             self._save(old_filename, metadata, settings)
         # Rename files
         if settings["rename_files"] or settings["move_files"]:
@@ -166,12 +167,16 @@ class File(LockableObject, Item):
         if settings["move_files"] and settings["move_additional_files"]:
             self._move_additional_files(old_filename, new_filename,
                                         settings)
+        if settings["copy_files"]: # save tages only to copied files (leave src files untouched )
+            self._save(new_filename, metadata, settings)
+
         # Delete empty directories
-        if settings["delete_empty_dirs"]:
+        if settings["delete_empty_dirs"] and not settings["copy_files"]:
             try:
                 os.removedirs(encode_filename(os.path.dirname(old_filename)))
             except EnvironmentError:
                 pass
+            print "deleting"
         # Save cover art images
         if settings["save_images_to_files"]:
             self._save_images(new_filename, metadata, settings)
@@ -269,8 +274,12 @@ class File(LockableObject, Item):
                 new_filename = "%s (%d)" % (tmp_filename, i)
                 i += 1
             new_filename = new_filename + ext
-            self.log.debug("Moving file %r => %r", old_filename, new_filename)
-            shutil.move(encode_filename(old_filename), encode_filename(new_filename))
+            if settings["copy_files"]:
+               self.log.debug("Copying file %r => %r", old_filename, new_filename)
+               shutil.copy(encode_filename(old_filename), encode_filename(new_filename))
+            else:
+               self.log.debug("Moving file %r => %r", old_filename, new_filename)
+               shutil.move(encode_filename(old_filename), encode_filename(new_filename))
             return new_filename
         else:
             return old_filename
@@ -327,7 +336,10 @@ class File(LockableObject, Item):
                     self.log.debug("File loaded in the tagger, not moving %r", old_file)
                     continue
                 self.log.debug("Moving %r to %r", old_file, new_file)
-                shutil.move(old_file, new_file)
+                if settings["copy_files"]:
+                   shutil.copy(old_file, new_file)
+                else:
+                   shutil.move(old_file, new_file)
 
     def remove(self, from_parent=True):
         if from_parent and self.parent:
